% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_etape.R
\name{cps}
\alias{cps}
\title{calcul de la quotité finale après application de la cps}
\usage{
cps(q, t)
}
\arguments{
\item{q}{quotient théorique en nombre entier (0.9 pour 90\%)}

\item{t}{taux de cps en nombre entier (0.01 pour 1\%)}
}
\value{
le quotient final après soustraction de la cps
}
\description{
calcul de la quotité finale après application de la cps
}
\examples{
cps(0.37, 0.06)
}
