% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_etape.R
\name{part}
\alias{part}
\title{Calcul de part pour un nombre donné de parents et d'enfants d'une famille}
\usage{
part(p, e)
}
\arguments{
\item{p}{nombre de parents (2 pour famille biparentale, 1 pour famille monoparentale)}

\item{e}{nombre d'enfants}
}
\value{
les nombre de parts de la famille
}
\description{
Calcul de part pour un nombre donné de parents et d'enfants d'une famille
}
\examples{
part(2, 3)
}
