% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_etape.R
\name{ponderation}
\alias{ponderation}
\title{pondération d'un quotient familial, la pondération converti le quotient familial en dénomination locale en EUR selon le taux de chancelerie en tennant compte des differences de pouvoir d'achat entre le lieux de vie de la famille et Paris.}
\usage{
ponderation(q, i, tx)
}
\arguments{
\item{q}{quotient familial en denomination locale}

\item{i}{index de parité de pouvoir d'achat (IPPA)}

\item{tx}{taux de chancellerie}
}
\value{
un nombre, le quotient familial ponderé en EUR
}
\description{
pondération d'un quotient familial, la pondération converti le quotient familial en dénomination locale en EUR selon le taux de chancelerie en tennant compte des differences de pouvoir d'achat entre le lieux de vie de la famille et Paris.
}
\examples{
ponderation(8500, 97, 1.23)
}
