% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotient.R
\name{quotient}
\alias{quotient}
\title{calcule le quotient d'un point x entre deux bornes a et b,
c'est a dire x est a q\% de b partant de a
en formulex = a + quotient * (b -a)}
\usage{
quotient(x, a, b)
}
\arguments{
\item{x}{point dont on veut connaitre le quotient}

\item{a}{borne basse}

\item{b}{borne haute}
}
\value{
a number
}
\description{
calcule le quotient d'un point x entre deux bornes a et b,
c'est a dire x est a q\% de b partant de a
en formulex = a + quotient * (b -a)
}
