% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_bourse.R
\name{quotient_final}
\alias{quotient_final}
\title{Calcul du quotient final des bourse AEFE en consideration de}
\usage{
quotient_final(salaire, scolarite, part, ippa, tx, t_cps)
}
\arguments{
\item{salaire}{salaire net en denomination locale (ex. 3 000 000 de Yen)}

\item{scolarite}{frais de scolarité strict en denomination locale (ex. 1 000 000 de Yen)}

\item{part}{part de la famille: 0.5 par enfants, 2 pour deux parents, 1.5 pour 1 parents seul (ex. 3 pour 2 parents et 2 enfants)}

\item{ippa}{taux IPPA (index de parité de pouvoir d'achat) fixé par l'AEFE pour l'école(ex. 78)}

\item{tx}{taux d'echange de la denomination locale en EUR, par Ex}

\item{t_cps}{CPS fixé par l'AEFE pour réduire les bourses à la hauteur du budget alloué}
}
\value{
le quotient qui détermine la hauteur de bourse accordé: (ex. 0.3, donc 30\% des frais seront pris en charge par la bourse)
}
\description{
Calcul du quotient final des bourse AEFE en consideration de
}
\examples{
quotient_final(7000000, 1100000, 2.5, 73, 0.0063, 0.05)

}
