% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_bourse_inv.R
\name{quotient_imp}
\alias{quotient_imp}
\title{Quotient implicite
valeur du quotient pondéré necéssaire pour obtenir la quotité théorique donné}
\usage{
quotient_imp(q)
}
\arguments{
\item{q}{quotité théorique}
}
\value{
un nombre, le quotient pondéré necessaire pour obtenir quotité theorique donné
}
\description{
Quotient implicite
valeur du quotient pondéré necéssaire pour obtenir la quotité théorique donné
}
\examples{
quotient_imp(0.75)
}
