% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_bourse.R
\name{quotient_theorique}
\alias{quotient_theorique}
\title{Calcul du quotient théorique, le pourcentage de bourse correspondant au besoin du requerreur avant ajustement budgétaire}
\usage{
quotient_theorique(salaire, scolarite, part, ippa, tx)
}
\arguments{
\item{salaire}{salaire net en denomination locale (ex. 3 000 000 de Yen)}

\item{scolarite}{frais de scolarité strict en denomination locale (ex. 1 000 000 de Yen)}

\item{part}{part de la famille: 0.5 par enfants, 2 pour deux parents, 1.5 pour 1 parents seul (ex. 3 pour 2 parents et 2 enfants)}

\item{ippa}{taux IPPA (index de parité de pouvoir d'achat) fixé par l'AEFE pour l'école(ex. 78)}

\item{tx}{taux d'echange de la denomination locale en EUR, par Ex}
}
\value{
le quotient qui détermine la hauteur de bourse correspondant au besoin reconnu par l'administration
}
\description{
Calcul du quotient théorique, le pourcentage de bourse correspondant au besoin du requerreur avant ajustement budgétaire
}
\examples{
 quotient_theorique(7000000, 1100000, 2.5, 73, 0.0063)
}
