% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_etape.R
\name{quotite}
\alias{quotite}
\title{quotité théorique de bourse}
\usage{
quotite(q)
}
\arguments{
\item{q}{quotient familial pondéré en EUR}
}
\value{
la quotité théorique de bourse
}
\description{
quotité théorique de bourse
}
\examples{
quotite(7000)
}
