% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_bourse.R
\name{quotite_post_cps}
\alias{quotite_post_cps}
\alias{quotite_theorique}
\alias{quotient_pondere}
\alias{quotient_f}
\title{Calcul du quotient final des bourse AEFE en consideration de}
\usage{
quotite_post_cps(salaire, scolarite, part, ippa, tx, t_cps)

quotite_theorique(salaire, scolarite, part, ippa, tx)

quotient_pondere(salaire, scolarite, part, ippa, tx)

quotient_f(salaire, scolarite, part)
}
\arguments{
\item{salaire}{salaire net en denomination locale (ex. 3 000 000 de Yen)}

\item{scolarite}{frais de scolarité strict en denomination locale (ex. 1 000 000 de Yen)}

\item{part}{part de la famille: 0.5 par enfants, 2 pour deux parents, 1.5 pour 1 parents seul (ex. 3 pour 2 parents et 2 enfants)}

\item{ippa}{taux IPPA (index de parité de pouvoir d'achat) fixé par l'AEFE pour l'école(ex. 78)}

\item{tx}{taux d'echange de la denomination locale en EUR, par Ex}

\item{t_cps}{taux CPS fixé par l'AEFE pour réduire les bourses à la hauteur du budget alloué}
}
\value{
la quotité qui détermine la hauteur de bourse accordée: (ex. 0.3, donc 30\% des frais seront pris en charge par la bourse)

la quotité qui détermine la hauteur de bourse correspondant au besoin reconnu par l'administration

un nombre, le quotient familial apres pondération en EUR

quotient familial en denomination locale
}
\description{
Calcul du quotient final des bourse AEFE en consideration de

Calcul du quotient théorique, le pourcentage de bourse correspondant au besoin de la famille avant ajustement budgétaire

Quotient ponderé

quotient familial
}
\examples{
quotite_post_cps(7000000, 1100000, 2.5, 73, 0.0063, 0.05)

 quotite_theorique(7000000, 1100000, 2.5, 73, 0.0063)
quotient_pondere(7000000, 1100000, 2.5, 73, 0.0063)
quotient_f(7000000, 1100000, 2.5)
}
