% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_bourse_inv.R
\name{quotite_theorique_imp}
\alias{quotite_theorique_imp}
\alias{quotient_pondere_imp}
\alias{quotient_familial_imp}
\alias{revenu_ref_imp}
\alias{revenu_net_imp}
\title{Quotite théorique implicite
valeur de quotité théorique qui resulterait dans la quotité finale donné, en fonction du taux de CPS applicable}
\usage{
quotite_theorique_imp(q, t)

quotient_pondere_imp(q, t)

quotient_familial_imp(q, i, tx, t)

revenu_ref_imp(q, p, i, tx, t)

revenu_net_imp(q, s, p, i, tx, t)
}
\arguments{
\item{q}{quotité finale de bourse accordée: (ex. 0.3, donc 30\% des frais seront pris en charge par la bourse)}

\item{t}{taux CPS fixé par l'AEFE pour réduire les bourses à la hauteur du budget alloué}

\item{i}{index de parité de pouvoir d'achat (IPPA) fixé par l'AEFE pour l'école(ex. 78)}

\item{tx}{taux d'echange de la denomination locale en EUR, par Ex}

\item{p}{part de la famille: 0.5 par enfants, 2 pour deux parents, 1.5 pour 1 parents seul (ex. 3 pour 2 parents et 2 enfants)}

\item{s}{frais de scolarité strict en denomination locale (ex. 1 000 000 de Yen)}
}
\value{
la quotite theorique de bourse

un nombre, les revenus net necessaire pour obternir la quotité finale dans le contexte parametré
}
\description{
Quotite théorique implicite
valeur de quotité théorique qui resulterait dans la quotité finale donné, en fonction du taux de CPS applicable

Revenu net implicite
}
\examples{
quotite_theorique_imp(0.3, 0.05)
revenu_net_imp(0.5, 1000, 3, 80, 1, 0.05)
}
